zomato_clone/requirements.txt
Django==6.0.2
Pillow==10.1.0
requests==2.31.0
